+++ home/urls.py
@@ -6,5 +6,7 @@ urlpatterns = [
     path('index/', views.index, name="index"),
     path("upload/",views.upload_data_view,name="upload_data"),
     # path("requestml/",views.requestmlresult),
+    path("requestnttu",views.requestnttu),
+    path("requesttai",views.requesttaipower),
 #
 ]
